+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<ma
 ## [v1.1.3]
 ### Fixed
 * Add missing classifier test cases for POSONLYARGS
+* Re-add the `tree` argument to the checker so flake8 identifies the plugin as needing to run
 
 ## [v1.1.2]
 ### Changed